minor, fix up strtoarri() so that the 0-th array member isn't checked

+++ lib/strops.h
@@ -86,9 +86,9 @@ extern size_t __ordtostr(char *buf, size_t bsz);
 
 /**
  * Take a string S, (case-insensitively) compare it to an array of strings ARR
- * of size NARR and return its index if found or -1 if not.
+ * of size NARR and return its index if found or 0 if not.
  * If S could be found in the array, point to the end of the string S in EP. */
-extern int32_t
+extern size_t
 strtoarri(const char *s, const char **ep, const char *const *arr, size_t narr);
 
 /**
